requirements.txt
asgiref==3.3.1
Django==3.1.3
djangorestframework==3.12.2
joblib==0.17.0
numpy==1.19.4
pytz==2020.4
scikit-learn==0.23.2
scipy==1.5.4
sqlparse==0.4.1
threadpoolctl==2.1.0
